Revert the fix for `url.port`, now addressed in #339

+++ code/go/ecs/url.go
@@ -45,7 +45,7 @@ type Url struct {
 	Domain string `ecs:"domain"`
 
 	// Port of the request, such as 443.
-	Port int64 `ecs:"port"`
+	Port int32 `ecs:"port"`
 
 	// Path of the request, such as "/search".
 	Path string `ecs:"path"`
